Move modified Jaylebreak's libraries to external files
To the main lua file more readable
